+++ packages/typespec-go/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.11.3 (unreleased)
+## 0.11.3 (2026-05-12)
 
 ### Bugs Fixed